@@ -22,24 +22,24 @@
   "dependencies": {
     "@agm/core": "^1.1.0",
     "@agm/snazzy-info-window": "^1.1.0",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/fire": "^5.3.0",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@types/jest": "^24.9.0",
+    "@angular/animations": "^9.0.2",
+    "@angular/cdk": "^9.0.1",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/fire": "^5.4.2",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "^9.0.2",
+    "@angular/material": "^9.0.1",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
+    "@types/jest": "^25.1.3",
     "angular2-text-mask": "^9.0.0",
     "echarts": "^4.6.0",
     "file-saver": "^2.0.2",
-    "firebase": "^7.7.0",
-    "firebase-tools": "^7.12.1",
+    "firebase": "^7.9.1",
+    "firebase-tools": "^7.13.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "howler": "^2.1.3",
@@ -52,37 +52,36 @@
     "ngx-echarts": "^4.2.2",
     "ngx-strongly-typed-forms": "^8.0.1",
     "rxjs": "^6.5.4",
-    "sass-loader": "7.0.1",
-    "saturn-datepicker": "^8.0.1",
+    "sass-loader": "^8.0.2",
     "stacktrace-js": "^2.0.2",
     "style-loader": "^1.1.3",
-    "tslib": "^1.11.0",
-    "xlsx": "^0.15.4",
-    "zone.js": "^0.10.2"
+    "tslib": "^1.10.0",
+    "xlsx": "^0.15.5",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular/cli": "~8.3.25",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular/cli": "~9.0.3",
+    "@angular/compiler-cli": "~9.0.2",
+    "@angular/language-service": "~9.0.2",
+    "@types/jasmine": "^3.5.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.149",
-    "@types/node": "~8.9.4",
+    "@types/node": "^13.7.4",
     "bootstrap": "^4.4.1",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
+    "codelyzer": "^5.2.1",
+    "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "node-sass": "4.13.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "^3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "node-sass": "^4.13.1",
+    "protractor": "^5.4.3",
+    "ts-node": "^8.6.2",
+    "tslint": "^6.0.0",
+    "typescript": "3.7.5",
     "webpack-bundle-analyzer": "^3.6.0"
   }
 }
