@@ -1,15 +1,14 @@
-import {Component, OnDestroy, OnInit} from '@angular/core';
-import {MatDialog} from '@angular/material';
-
-import {ConfirmDialogComponent} from 'app/components/confirm-dialog/confirm-dialog.component';
-import {AdminConfigService} from 'app/services/core/admin-config.service';
-import {CoreService} from 'app/services/core/core.service';
-import {ReplaySubject} from 'rxjs';
-import {takeUntil} from 'rxjs/operators';
-import {emptyomc, OMC} from '../../../../models/Daudi/omc/OMC';
-import {deepCopy} from '../../../../models/utils/deepCopy';
-import {NotificationService} from '../../../../shared/services/notification.service';
-import {CommunicationService} from '../../communication.service';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from 'app/components/confirm-dialog/confirm-dialog.component';
+import { AdminConfigService } from 'app/services/core/admin-config.service';
+import { CoreService } from 'app/services/core/core.service';
+import { ReplaySubject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { emptyomc, OMC } from '../../../../models/Daudi/omc/OMC';
+import { deepCopy } from '../../../../models/utils/deepCopy';
+import { NotificationService } from '../../../../shared/services/notification.service';
+import { CommunicationService } from '../../communication.service';
 
 @Component({
   selector: 'app-company',
@@ -20,8 +19,8 @@ export class ConfigComponent implements OnInit, OnDestroy {
   /**
    * the payment channels that require collection of transaction details
    */
-    // customizablepaymentchannels: Array<PaymentChannel>;
-    // confirmDialogRef: MatDialogRef<FuseConfirmDialogComponent>;
+  // customizablepaymentchannels: Array<PaymentChannel>;
+  // confirmDialogRef: MatDialogRef<FuseConfirmDialogComponent>;
   defaultlat = -1.3373943;
   defaultlng = 36.7208522;
   zoom = 15;
