@@ -1,15 +1,15 @@
-import {Component, OnDestroy, OnInit} from '@angular/core';
-import {AngularFireFunctions} from '@angular/fire/functions';
-import {MatDialog} from '@angular/material';
-import {MapsComponent} from 'app/components/maps/maps.component';
-import {CoreService} from 'app/services/core/core.service';
-import {ReplaySubject} from 'rxjs';
-import {takeUntil} from 'rxjs/operators';
-import {SyncRequest} from '../../../../models/Cloud/Sync';
-import {Depot, emptydepot} from '../../../../models/Daudi/depot/Depot';
-import {emptyomc, OMC} from '../../../../models/Daudi/omc/OMC';
-import {MyTimestamp} from '../../../../models/firestore/firestoreTypes';
-import {NotificationService} from '../../../../shared/services/notification.service';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { AngularFireFunctions } from '@angular/fire/functions';
+import { MapsComponent } from 'app/components/maps/maps.component';
+import { CoreService } from 'app/services/core/core.service';
+import { ReplaySubject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { SyncRequest } from '../../../../models/Cloud/Sync';
+import { Depot, emptydepot } from '../../../../models/Daudi/depot/Depot';
+import { emptyomc, OMC } from '../../../../models/Daudi/omc/OMC';
+import { MyTimestamp } from '../../../../models/firestore/firestoreTypes';
+import { NotificationService } from '../../../../shared/services/notification.service';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'depot-management',
@@ -26,7 +26,7 @@ export class DepotManagementComponent implements OnInit, OnDestroy {
   zoom = 7;
   creatingsync = false;
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
-  company: OMC = {...emptyomc};
+  company: OMC = { ...emptyomc };
 
   constructor(
     private notification: NotificationService,
