@@ -1,7 +1,6 @@
 import { AgmCoreModule } from "@agm/core";
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
-import { MatSidenavModule } from "@angular/material";
 import { SharedModule } from "app/shared/shared.module";
 import { AdminRoutingModule } from "./admin-routing.module";
 import { AdminComponent } from "./admin.component";
@@ -20,6 +19,7 @@ import { PaymentsComponent } from "./pages/payments/payments.component";
 import { SmsLogsComponent } from "./pages/sms-logs/sms-logs.component";
 import { UserLevelsComponent } from "./pages/user-levels/user-levels.component";
 import { UserManagementComponent } from "./pages/user-management/user-management.component";
+import { MatSidenavModule } from '@angular/material/sidenav';
 
 @NgModule({
   declarations: [
