@@ -1,7 +1,6 @@
 import { SelectionModel } from '@angular/cdk/collections';
 import { Component, Inject, OnDestroy, OnInit, Optional, ViewChild } from '@angular/core';
 import { AngularFireFunctions } from '@angular/fire/functions';
-import { MAT_DIALOG_DATA, MatDialog, MatDialogRef, MatPaginator, MatSnackBar, MatSort, MatTableDataSource } from '@angular/material';
 import { SendMsgComponent } from 'app/components/send-msg/send-msg.component';
 import { AdminService } from 'app/services/core/admin.service';
 import { CoreService } from 'app/services/core/core.service';
@@ -14,6 +13,11 @@ import { MyTimestamp } from '../../../../models/firestore/firestoreTypes';
 import { NotificationService } from '../../../../shared/services/notification.service';
 import { CoreAdminService } from '../../services/core.service';
 import { CompanyMembersComponent } from '../company-members/company-members.component';
+import { MatTableDataSource } from '@angular/material/table';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 @Component({
   selector: 'customer-management',
@@ -32,6 +36,7 @@ export class CustomerManagementComponent implements OnInit, OnDestroy {
   selection = new SelectionModel<DaudiCustomer>(true, []);
   loadingcompanies = true;
   savingcompany = false;
+  typedValue: string
 
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
 
