@@ -1,6 +1,5 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { MatSnackBar } from '@angular/material';
 import { Router } from '@angular/router';
 import { CoreService } from 'app/services/core/core.service';
 import { DepotService } from 'app/services/core/depot.service';
@@ -21,6 +20,8 @@ import { FuelBoundstats } from '../charts/charts.config';
 import { singleFuelpricestat } from '../charts/prices';
 import { fuelgauge } from '../charts/qty';
 import { saleStats } from '../charts/sales';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
 
 @Component({
   selector: 'app-stats',
